okex-python-sdk-api/requirements.txt
certifi==2018.11.29
chardet==3.0.4
idna==2.8
psycopg2==2.7.6.1
pycoin==0.80
python-redis-lock==3.3.1
pytz==2018.9
PyYAML==3.13
redis==2.10.6
requests==2.21.0
urllib3==1.24.1
#
base58==0.2.1
eth-keys==0.2.1
eth-utils==1.4.1
mnemonic==0.13
pycryptodome==3.8.1
rlp==0.6.0
two1==3.10.8
psycopg2-binary==2.8.2
web3==5.11.1
eth-keys==0.2.1
eth-utils==1.4.1
sqlalchemy==1.2.7
celery==3.1.25
dataclasses==0.8

flask==1.1.2
